ft(ride): Join ride
- Add join ride component

[Finishes #166997190]
