+++ src/App.jsx
@@ -1,8 +1,10 @@
 import React, { Component, Fragment } from 'react';
-import { BrowserRouter, Route, Switch } from 'react-router-dom';
+import { Route, Switch, withRouter } from 'react-router-dom';
+import PropTypes from 'prop-types';
 import Home from './component/home/Home';
 import Header from './component/header/Header';
 import CreateRideOffer from './component/rideOffer/CreateRideOffer';
+import Footer from './component/footer/Footer';
 import './styles/style.css';
 
 class App extends Component {
@@ -11,21 +13,30 @@ class App extends Component {
   };
 
   render() {
-    const { isAuth } = this.state;
+    const {
+      location: { pathname },
+    } = this.props;
+
     return (
-      <div className="container">
-        <BrowserRouter>
-          {!isAuth && <Header />}
-          <Fragment>
+      <Fragment>
+        <div className="container">
+          {pathname !== '/' && <Header />}
+          <div>
             <Switch>
               <Route exact path="/" component={Home} />
               <Route path="/ride-offer" component={CreateRideOffer} />
               <Route path="/request" component={CreateRideOffer} />
             </Switch>
-          </Fragment>
-        </BrowserRouter>
-      </div>
+          </div>
+          {pathname !== '/' && <Footer />}
+        </div>
+      </Fragment>
     );
   }
 }
-export default App;
+App.propTypes = {
+  location: PropTypes.shape({
+    pathname: PropTypes.string,
+  }).isRequired,
+};
+export default withRouter(App);
