+++ src/component/commons/common.scss
@@ -2,3 +2,4 @@ $white: #ffffff;
 $red: #ff0000;
 $gray: rgb(128, 128, 128);
 $green_dark: #339933;
+$black: #000000;
