+++ src/component/footer/Footer.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import './Footer.scss';
+
+const Footer = () => (
+  <div className="row footer-container">
+    <div className="col-12 footer">
+      <span>&#169;</span>
+      {' '}
+      <span>Okoro Emeka</span>
+    </div>
+  </div>
+);
+
+export default Footer;
