+++ src/component/footer/Footer.scss
@@ -1,0 +1,11 @@
+@import '../commons/common.scss';
+.footer-container {
+  padding: 0px;
+  .footer {
+    height: 80px;
+    padding: 0 5px;
+    color: $white;
+    background-color: $black;
+    margin-top: 5%;
+  }
+}
