+++ src/component/home/Home.scss
@@ -2,9 +2,9 @@
   .row {
     padding: 0;
     #banner {
+      height: 100vh;
       color: white;
       text-align: center;
-      height: 100vh;
       background-image: url('../../assets/images/taxi.jpg');
       background-position: center;
       background-repeat: no-repeat;
